Poll for tx status update
